🚩 Community 댓글창 개수 및 디자인 수정
