+++ src/Page/Community/scss/community.module.scss
@@ -267,6 +267,10 @@
           }
         }
 
+        .inputGroup.hidden {
+          display: none;
+        }
+
         .inputGroup {
           height: 52px;
           margin: 1rem 1rem 1rem 1rem;
